Add .catch(() => {}) to all modal.result.open promise chains
