--- app-frontend/src/app/components/common/feedbackModal/feedbackModal.component.js
@@ -1,15 +1,0 @@
-// Component code
-import feedbackModalTpl from './feedbackModal.html';
-
-const rfFeedbackModal = {
-    templateUrl: feedbackModalTpl,
-    bindings: {
-        close: '&',
-        dismiss: '&',
-        modalInstance: '<',
-        resolve: '<'
-    },
-    replace: true
-};
-
-export default rfFeedbackModal;
