+++ app-frontend/src/app/components/common/feedbackModal/feedbackModal.module.js
@@ -1,5 +1,17 @@
 import angular from 'angular';
-import FeedbackModalComponent from './feedbackModal.component.js';
+
+import feedbackModalTpl from './feedbackModal.html';
+
+const FeedbackModalComponent = {
+    templateUrl: feedbackModalTpl,
+    bindings: {
+        close: '&',
+        dismiss: '&',
+        modalInstance: '<',
+        resolve: '<'
+    },
+    replace: true
+};
 
 const FeedbackModalModule = angular.module('components.common.feedbackModal', []);
