+++ app-frontend/src/app/components/datasources/datasourceItem/datasourceItem.module.js
@@ -60,7 +60,7 @@ class DatasourceItemController {
             }, (err) => {
                 this.$log.debug('error deleting datasource', err);
             });
-        });
+        }).catch(() => {});
     }
 }
