+++ app-frontend/src/app/components/histogram/nodeHistogram/nodeHistogram.module.js
@@ -436,7 +436,7 @@ class NodeHistogramController {
                     baseScheme
                 })
             });
-        });
+        }).catch(() => {});
     }
 }
