+++ app-frontend/src/app/components/lab/classifyNode/classifyNode.module.js
@@ -103,7 +103,7 @@ class ClassifyNodeController {
                 }),
                 hard: true
             });
-        });
+        }).catch(() => {});
     }
 }
