+++ app-frontend/src/app/pages/imports/raster/raster.module.js
@@ -35,7 +35,7 @@ class RasterListController {
 
         this.activeModal.result.then(() => {
             this.checkPendingImports();
-        });
+        }).catch(() => {});
     }
 
     checkPendingImports() {
