+++ app-frontend/src/app/pages/projects/edit/aoi-parameters/aoi-parameters.module.js
@@ -191,7 +191,7 @@ class AOIParametersController {
                 }
             }).result.then(selectedDay => {
                 this.updateStartDate(selectedDay);
-            });
+            }).catch(() => {});
     }
 
     updateFrequency(ms) {
