+++ app-frontend/src/app/pages/user/settings/api/api.html
@@ -20,10 +20,10 @@
       </div>
     </form>
     <div class="list-group"
-         ng-show="$ctrl.createError">
+         ng-show="$ctrl.tokenCreateError">
       <span class="list-placeholder"
             style="color: red; font-weight: bold;">
-        There was an error generating a new token
+        There was an error generating a new token.
       </span>
     </div>
     <div class="list-group">
