+++ app-frontend/src/app/pages/user/settings/map/map.controller.js
@@ -62,7 +62,7 @@ class MapTokensController {
                     this.fetchTokens();
                 }
             );
-        });
+        }).catch(() => {});
     }
 
     updateToken(token, name) {
