+++ app-frontend/src/app/services/scenes/planetRepository.service.js
@@ -48,6 +48,8 @@ export default (app) => {
                             this.$state.go('user.settings.connections');
                         }, () => {
                             reject();
+                        }).catch(() => {
+                            reject();
                         });
                     }
                 }, () => {
